Add sheet view as its own file (delete scan view). Remove disconnecting scan state.
