+++ App/Biodome/Biodome/Views/ContentView.swift
@@ -9,66 +9,7 @@ import SwiftUI
 
 // Put Name variable somehwere ??? Figure out name and UUID stuff some more -> add info tab
 // Connection timeout / error message?
-// Split this into 2 files?
 // If removed from list is peripheral refound?
-// Add progress wheel on connected device while connecting
-// Conencting / disconnecting logic messages might be a little out of sync
-
-struct SelectedRow: View {
-    @EnvironmentObject var bleManager: BLEManager
-    var peripheral: Peripheral
-    
-    var body: some View {
-        HStack {
-            Text(peripheral.name)
-            Spacer()
-            Text(bleManager.connectionState).foregroundColor(.secondary)
-        }
-    }
-}
-
-
-struct PeripheralRow: View {
-    @EnvironmentObject var bleManager: BLEManager
-    var peripheral: Peripheral
-    
-    var body: some View {
-        HStack {
-            Text(peripheral.name)
-            Spacer()
-            Button(action: {
-                bleManager.connect(selected: peripheral)
-            }){}
-        }
-    }
-}
-
-
-struct SheetView: View {
-    @Environment(\.dismiss) var dismiss
-    @EnvironmentObject var bleManager: BLEManager
-    
-    var body: some View {
-        
-        List {
-            Section(header: Text("Connected Device")){
-                ForEach(bleManager.connectionSelect) { peripheral in
-                    SelectedRow(peripheral: peripheral)
-                }
-            }
-            
-            Section(header:
-                HStack{
-                    Text("Devices  ")
-                    ProgressView()
-                }) {
-                ForEach(bleManager.peripherals) { peripheral in
-                    PeripheralRow(peripheral: peripheral)
-                }
-            }
-        }.listStyle(InsetGroupedListStyle())
-    }
-}
 
 
 struct ContentView: View {
